+++ external-types.d.ts
@@ -1,1 +1,5 @@
 declare module '*.scss';
+declare module 'postcss-rtl';
+declare module 'postcss-exclude-files';
+declare module 'rollup-plugin-babel';
+declare module 'deep-freeze';