+++ packages/plugin-file-upload/web/src/icons/getIcon.tsx
@@ -6,6 +6,7 @@ import otherIcon from './otherIcon';
 import imageIcon from './imageIcon';
 import folderIcon from './folderIcon';
 import docIcon from './docIcon';
+import { ComponentType } from 'react';
 
 const doc = new RegExp('^doc([a-z]*)$');
 const excel = new RegExp('^xl([a-z]*)$');
@@ -61,7 +62,11 @@ const iconList = [
   { data: folder, icon: folderIcon },
 ];
 
-function getIconFromList(type, checkList, typePredicate) {
+function getIconFromList(
+  type: string,
+  checkList: { data: string[] | RegExp; icon: ComponentType }[],
+  typePredicate: (testSet: RegExp | string[], type: string) => boolean
+) {
   let retVal;
   checkList.some(({ data, icon }) => {
     if (typePredicate(data, type)) {
@@ -73,10 +78,18 @@ function getIconFromList(type, checkList, typePredicate) {
   return retVal;
 }
 
-export const getIcon = type => {
+export const getIcon = (type: string) => {
   const icon =
     type &&
-    (getIconFromList(type, iconRegList, (regExp, type) => regExp.test(type)) ||
-      getIconFromList(type, iconList, (typeList, type) => typeList.some(e => e === type)));
+    (getIconFromList(
+      type,
+      iconRegList as { data: RegExp; icon: ComponentType }[],
+      (regExp: RegExp, type) => regExp.test(type)
+    ) ||
+      getIconFromList(
+        type,
+        iconList as { data: string[]; icon: ComponentType }[],
+        (typeList: string[], type) => typeList.some(e => e === type)
+      ));
   return icon || otherIcon;
 };