+++ packages/ricos-content/web/src/version/versioningUtils.ts
@@ -12,21 +12,14 @@ const toVersion = (versionString: string) => {
     .slice(0, 3);
 };
 
-/**
- * compareVersions
- * @description compares 2 version strings. assumption: versions are valid and exact (semantics not supported)
- * @param {string} left version string "major.minor.revision"
- * @param {string} right version string "major.minor.revision"
- * @returns integer: left > right => 1, left === right => 0, left < right => -1
- */
-const compareVersions = (left: string, right: string) => {
+const compareVersions: (left: string, right: string) => -1 | 0 | 1 = (left, right) => {
   const leftVersion = toVersion(left);
   const rightVersion = toVersion(right);
-  const diff = leftVersion.map((left, idx) => Math.sign(left - rightVersion[idx]));
+  const diff = leftVersion.map((left, idx) => Math.sign(left - rightVersion[idx]) as -1 | 0 | 1);
   return diff.find(num => num !== 0) || 0;
 };
 
-const evaluations = {
+const evaluations: { [operator: string]: (left: string, right: string) => boolean } = {
   '<': (left: string, right: string) => compareVersions(left, right) < 0,
   '>': (left: string, right: string) => compareVersions(left, right) > 0,
   '=': (left: string, right: string) => compareVersions(left, right) === 0,
@@ -34,14 +27,7 @@ const evaluations = {
   '>=': (left: string, right: string) => compareVersions(left, right) >= 0,
 };
 
-/**
- * evaluateVersion
- * @description evaluates version string with conditional semver
- * @param {string} tested version to evaluate
- * @param {string} semver consisting of operator and version (<1.0.0, >=2.3.3). supports < > = operators
- * @returns {boolean} true if version meets the semver condition
- */
-const evaluateVersion = (tested, semver) => {
+const evaluateVersion = (tested: string, semver: string) => {
   const [, operator, version] = /([<>=]{0,2})(.*)/gm.exec(semver) as string[];
   return evaluations[operator || '='](tested, version);
 };